requirements.txt
fastapi[standard]
httpx
beautifulsoup4
supabase
python-dotenv
telethon
tavily-python
